Refactor FocusWatcher into its own class
